+++ apps/router-e2e/app.config.js
@@ -84,7 +84,7 @@ module.exports = {
             : undefined,
         unstable_useServerDataLoaders: process.env.E2E_ROUTER_SERVER_LOADERS === 'true',
         unstable_useServerMiddleware: process.env.E2E_ROUTER_SERVER_MIDDLEWARE === 'true',
-        unstable_splitView: true
+        unstable_splitView: process.env.E2E_ROUTER_SPLIT_VIEW === 'true',
       },
     ],
   ],