+++ packages/expo-router/src/split-view/split-view.tsx
@@ -1,9 +1,6 @@
+import Constants from 'expo-constants';
 import React, { createContext, isValidElement, use, type ReactNode } from 'react';
-import {
-  SplitViewHost,
-  SplitViewScreen,
-  type SplitViewHostProps,
-} from 'react-native-screens/experimental';
+import { SplitViewHost, SplitViewScreen, type SplitViewHostProps } from 'react-native-screens';
 
 import { SplitViewColumn, SplitViewInspector } from './elements';
 import { IsWithinLayoutContext } from '../layouts/withLayoutContext';
@@ -23,6 +20,19 @@ function SplitViewNavigator({ children, ...splitViewHostProps }: SplitViewProps)
   // TODO: Add better way of detecting if SplitView is rendered inside Native navigator.
   if (use(IsWithinLayoutContext)) {
     throw new Error('SplitView cannot be used inside another navigator, except for Slot.');
+  }
+
+  if (!Constants.expoConfig?.extra?.router?.unstable_splitView) {
+    throw new Error(
+      'SplitView is not enabled. Make sure to enable it in your expo-router configuration with "unstable_splitView": true. After enabling, make sure to prebuild your app.'
+    );
+  }
+
+  if (process.env.EXPO_OS !== 'ios') {
+    console.warn(
+      'SplitView is only supported on iOS. The SplitView will behave like a Slot navigator on other platforms.'
+    );
+    return <Slot />;
   }
 
   const WrappedSlot = () => (